[UPDATE] - tools transference

+++ app/src/main/java/com/simaht/modules/dashboard_mh/AssignTool/Tool.kt
@@ -4,7 +4,7 @@ class Tool(var name : String,
            var brand : String,
            var activedOn : String,
            var seguro : Int,
-           var serialNumber: Int,
+           var serialNumber: Long,
            var price : String,
            var status : Boolean,
            var bill: String)
